+++ src/containers/CrudEndpointUi/layout.js
@@ -36,6 +36,7 @@ import ReasonListsForm from '../Form/ReasonLists';
 import QueuesForm from '../Form/Queues';
 import FlowsForm from '../Form/Flows';
 import CopyFlowForm from '../Form/Flows/Copy';
+import TransferListsForm from '../Form/TransferLists';
 //hygen-inject-before3
 
 // AddMembersToList table modal
@@ -58,6 +59,7 @@ import ReasonsDetailsPanel from '../SidePanelDetails/Reasons';
 import ReasonListsDetailsPanel from '../SidePanelDetails/ReasonLists';
 import QueuesDetailsPanel from '../SidePanelDetails/Queues';
 import FlowsDetailsPanel from '../SidePanelDetails/Flows';
+import TransferListsDetailsPanel from '../SidePanelDetails/TransferLists';
 //hygen-inject-before4
 
 const Wrapper = styled.div`
@@ -202,6 +204,14 @@ const createFormRoutes = [
     component: () => (
       <DetailsPanel>
         <FlowsForm />
+      </DetailsPanel>
+    )
+  },
+  {
+    path: '/configuration/transferLists',
+    component: () => (
+      <DetailsPanel>
+        <TransferListsForm />
       </DetailsPanel>
     )
   }
@@ -357,6 +367,16 @@ const detailsPanelRoutes = [
           <FlowsForm />
         </FlowsDetailsPanel>
       </DetailsPanel>
+    )
+  },
+  {
+    path: '/configuration/transferLists',
+    component: () => (
+      <NoScrollDetailsPanel>
+        <TransferListsDetailsPanel>
+          <TransferListsForm />
+        </TransferListsDetailsPanel>
+      </NoScrollDetailsPanel>
     )
   }
   //hygen-inject-before2