+++ src/containers/Form/TransferLists/layout.js
@@ -1,0 +1,44 @@
+/*
+ * Copyright © 2015-2019 Serenova, LLC. All rights reserved.
+ */
+
+/**
+ *
+ * TransferListsForm
+ *
+ */
+
+import React from 'react';
+import PropTypes from 'prop-types';
+import { DetailHeader, InputField } from 'cx-ui-components';
+
+export default function TransferListsForm({ handleSubmit, isSaving, userHasUpdatePermission, key }) {
+  return (
+    <form onSubmit={handleSubmit} key={key}>
+      <DetailHeader text="Details" />
+      <InputField
+        name="name"
+        label="Name *"
+        componentType="input"
+        inputType="text"
+        automation="transferListsNameInput"
+        disabled={isSaving || !userHasUpdatePermission}
+      />
+      <InputField
+        name="description"
+        label="Description"
+        componentType="textarea"
+        inputType="text"
+        automation="transferListsDescriptionInput"
+        disabled={isSaving || !userHasUpdatePermission}
+      />
+    </form>
+  );
+}
+
+TransferListsForm.propTypes = {
+  key: PropTypes.string,
+  handleSubmit: PropTypes.func.isRequired,
+  isSaving: PropTypes.bool,
+  userHasUpdatePermission: PropTypes.bool
+};