+++ src/containers/Form/TransferLists/validation.js
@@ -1,0 +1,9 @@
+/*
+ * Copyright © 2015-2019 Serenova, LLC. All rights reserved.
+ */
+
+import { isEmpty } from 'serenova-js-utils/strings';
+
+export const formValidation = values => ({
+  name: isEmpty(values.get('name')) && 'Please enter a name.'
+});