+++ src/containers/SidePanelDetails/TransferLists/__tests__/index.test.js
@@ -1,0 +1,35 @@
+/*
+ * Copyright © 2015-2019 Serenova, LLC. All rights reserved.
+ */
+
+import React from 'react';
+import { Map } from 'immutable';
+import { createStore } from 'redux';
+import { shallow } from 'enzyme';
+import TransferListsDetailsPanel from '../';
+import {
+  getSelectedEntity,
+  userHasUpdatePermission,
+  isInherited,
+  isSaving
+} from '../../../../redux/modules/entities/selectors';
+
+jest.mock('../../../../redux/modules/entities/selectors');
+getSelectedEntity.mockImplementation(
+  () =>
+    new Map({
+      id: 'mockId',
+      name: 'mockName',
+      description: 'mockDescription'
+    })
+);
+userHasUpdatePermission.mockImplementation(() => true);
+isInherited.mockImplementation(() => false);
+isSaving.mockImplementation(() => false);
+
+describe('TransferListsDetailsPanel Renders', () => {
+  it('renders', () => {
+    const store = createStore(state => state);
+    expect(shallow(<TransferListsDetailsPanel store={store}>Child</TransferListsDetailsPanel>)).toMatchSnapshot();
+  });
+});