+++ src/redux/modules/form/selectors.js
@@ -61,6 +61,13 @@ export const selectCreateUserFormInitialValues = state => {
   });
 };
 
+export const selectTransferListsCreateFormInitialValues = state => {
+  return new Map({
+    active: false,
+    endpoints: []
+  });
+};
+
 export const formSubmission = (values, dispatch, props) => dispatch(onFormSubmit(values, props));
 
 export const subEntityFormSubmission = (values, dispatch, props) => dispatch(onSubEntityFormSubmit(values, props));